added sfh to match/fake and fixed crowd argument in match_phot

+++ scripts/fake.py
@@ -1,12 +1,18 @@
+import argparse
+import os
+import sys
 from .fileio import read_calcsfh_param, fake_param_fmt
 from .sfh import SFH
 
 
-def make_fakeparam(param, sfhfile):
+def make_fakeparam(param, sfhfile, outfile=None, overwrite=False):
     """
     Convert a calcsfh solution and parameter file into a fake parameter file
     """
-    outfile = param.replace('param', 'fparam')
+    if outfile is None:
+        ext = os.path.splitext(sfhfile)[1]
+        outfile = sfhfile.replace(ext, '.fparam')
+
     pdict = read_calcsfh_param(param)
     sfh = SFH(sfhfile)
 
@@ -23,5 +29,37 @@ def make_fakeparam(param, sfhfile):
                                                           sfh.data.sfr[i],
                                                           sfh.data.mh[i])
                        for i in range(ntbins)])
-    with open(outfile, 'w') as f:
-        f.write(line)
+    if not os.path.isfile(outfile) or overwrite:
+        with open(outfile, 'w') as f:
+            f.write(line)
+        print('wrote {:s}'.format(outfile))
+    return
+
+
+def parse_args(argv=None):
+    parser = argparse.ArgumentParser(description=make_fakeparam.__doc__)
+
+    parser.add_argument('--overwrite', action='store_true',
+                        help='overwrite existing fake param file')
+
+    parser.add_argument('--outfile', default=None,
+                        help='filename to write to')
+
+    parser.add_argument('param', default=None,
+                        help='calcsfh input parameter file')
+
+    parser.add_argument('sfh', default=None,
+                        help='SFH file (output of zcombine)')
+
+    return parser.parse_args(argv)
+
+
+def main(argv=None):
+    args = parse_args(argv)
+    make_fakeparam(args.param, args.sfh, outfile=args.outfile,
+                   overwrite=args.overwrite)
+    return
+
+
+if __name__ == "__main__":
+    sys.exit(main())
